move files around and rename and split helpers for connect
