+++ shared/login/login/index.desktop.js
@@ -1,10 +1,11 @@
 // @flow
 import React, {Component} from 'react'
-import type {Props} from './index.render'
 import {Box, UserCard, Text, Button, FormWithCheckbox, Dropdown} from '../../common-adapters'
 import {globalStyles, globalColors} from '../../styles'
 
-class LoginRender extends Component<void, Props, void> {
+import type {Props} from '.'
+
+class Login extends Component<void, Props, void> {
   render() {
     const inputProps = {
       hintText: 'Passphrase',
@@ -75,4 +76,4 @@ const stylesContainer = {
   backgroundColor: globalColors.white,
 }
 
-export default LoginRender
+export default Login
