+++ shared/login/register/code-page/dumb.js
@@ -1,7 +1,8 @@
 // @flow
-import Render from './index.render'
+import Render from '.'
 import {qrGenerate} from '../../../actions/login/provision-helpers'
 import {codePageDeviceRoleExistingPhone, codePageDeviceRoleExistingComputer} from '../../../constants/login'
+
 import type {DumbComponentMap} from '../../../constants/types/more'
 
 const baseMock = {
