+++ shared/login/register/error/index.js.flow
@@ -7,4 +7,4 @@ export type Props = {
   onBack: () => void,
 }
 
-export default class Render extends Component<void, Props, void> {}
+export default class RenderError extends Component<void, Props, void> {}
