+++ shared/login/register/gpg-sign/index.native.js
