+++ shared/login/register/paper-key/container.js
@@ -1,15 +1,26 @@
 // @flow
 import React, {Component} from 'react'
-import RenderPaperKey from './index.render'
+import PaperKey from '.'
 import {connect} from 'react-redux'
 import * as Creators from '../../../actions/login/creators'
 import HiddenString from '../../../util/hidden-string'
 
 import type {RouteProps} from '../../../route-tree/render-route'
-import type {Props, State} from './index'
 import type {TypedState} from '../../../constants/reducer'
 
-class PaperKey extends Component<void, Props, State> {
+export type Props = {
+  onSubmit: (paperKey: string) => void,
+  onBack: () => void,
+  waitingForResponse?: boolean,
+  error: string,
+}
+
+export type State = {
+  paperKey: string,
+}
+
+// TODO remove this class
+class _PaperKey extends Component<void, Props, State> {
   state: State
 
   constructor(props) {
@@ -22,7 +33,7 @@ class PaperKey extends Component<void, Props, State> {
 
   render() {
     return (
-      <RenderPaperKey
+      <PaperKey
         onSubmit={() => this.props.onSubmit(this.state.paperKey)}
         onChangePaperKey={paperKey => this.setState({paperKey})}
         onBack={this.props.onBack}
@@ -41,14 +52,15 @@ type OwnProps = RouteProps<
   {}
 >
 
+const mapStateToProps = (state: TypedState, {routeProps: {error}}: OwnProps) => ({
+  waitingForResponse: state.engine.get('rpcWaitingStates').get('loginRpc'),
+  error,
+})
+
+const mapDispatchToProps = dispatch => ({
+  onBack: () => dispatch(Creators.onBack()),
+  onSubmit: paperkey => dispatch(Creators.submitPassphrase(new HiddenString(paperkey), false)),
+})
+
 // $FlowIssue
-export default connect(
-  (state: TypedState, {routeProps: {error}}: OwnProps) => ({
-    waitingForResponse: state.engine.get('rpcWaitingStates').get('loginRpc'),
-    error,
-  }),
-  dispatch => ({
-    onBack: () => dispatch(Creators.onBack()),
-    onSubmit: paperkey => dispatch(Creators.submitPassphrase(new HiddenString(paperkey), false)),
-  })
-)(PaperKey)
+export default connect(mapStateToProps, mapDispatchToProps)(_PaperKey)
