+++ shared/login/register/paper-key/index.js.flow
@@ -2,16 +2,12 @@
 import {Component} from 'react'
 
 export type Props = {
-  onSubmit: (paperKey: string) => void,
   onBack: () => void,
-  waitingForResponse?: boolean,
-  error: string,
-}
-
-export type State = {
+  onSubmit: () => void,
+  onChangePaperKey: (val: string) => void,
   paperKey: string,
+  error: string,
+  waitingForResponse?: ?boolean,
 }
 
-export default class PaperKey extends Component<void, Props, State> {
-  state: State
-}
+export default class PaperKey extends Component<void, Props, void> {}
