+++ shared/login/register/paper-key/index.native.js
@@ -1,11 +1,12 @@
 // @flow
 import Container from '../../forms/container'
 import React, {Component} from 'react'
-import type {Props} from './index.render'
 import {Button, Icon, Input, Text} from '../../../common-adapters'
 import {globalMargins} from '../../../styles'
 
-class PaperKeyRender extends Component<void, Props, void> {
+import type {Props} from '.'
+
+class PaperKey extends Component<void, Props, void> {
   render() {
     return (
       <Container style={stylesContainer} onBack={this.props.onBack}>
@@ -46,4 +47,4 @@ const stylesInput = {
   alignSelf: 'stretch',
 }
 
-export default PaperKeyRender
+export default PaperKey
