--- shared/login/register/paper-key/index.render.js.flow
@@ -1,13 +1,0 @@
-// @flow
-import {Component} from 'react'
-
-export type Props = {
-  onBack: () => void,
-  onSubmit: () => void,
-  onChangePaperKey: (val: string) => void,
-  paperKey: string,
-  error: string,
-  waitingForResponse?: ?boolean,
-}
-
-export default class PaperKey extends Component<void, Props, void> {}
