+++ shared/login/register/select-other-device/container.js
@@ -1,0 +1,23 @@
+// @flow
+import SelectOtherDevice from '.'
+import * as Creators from '../../../actions/login/creators'
+import {connect} from 'react-redux'
+
+import type {TypedState} from '../../../constants/reducer'
+import type {Device} from '../../../constants/types/more'
+
+type OwnProps = {
+  routeProps: {
+    devices: Array<Device>,
+  },
+}
+
+const mapStateToProps = (s: TypedState, {routeProps: {devices}}: OwnProps) => ({devices})
+const mapDispatchToProps = dispatch => ({
+  onBack: () => dispatch(Creators.onBack()),
+  onWont: () => dispatch(Creators.onWont()),
+  onSelect: deviceId => dispatch(Creators.selectDeviceId(deviceId)),
+})
+
+// $FlowIssue
+export default connect(mapStateToProps, mapDispatchToProps)(SelectOtherDevice)
