+++ shared/login/register/select-other-device/index.js.flow
@@ -9,4 +9,4 @@ export type Props = {
   onBack: () => void,
 }
 
-export default class Render extends Component<void, Props, void> {}
+export default class SelectOtherDevice extends Component<void, Props, void> {}
