+++ shared/login/register/username-or-email/index.js.flow
@@ -7,4 +7,4 @@ export type Props = {
   waitingForResponse: boolean,
 }
 
-export default class PaperKey extends Component<void, Props, void> {}
+export default class UsernameOrEmail extends Component<void, Props, void> {}
