+++ shared/login/register/username-or-email/index.native.js
@@ -1,13 +1,14 @@
 // @flow
-import Container from '../../forms/container.desktop'
+import Container from '../../forms/container'
 import React, {Component} from 'react'
-import type {Props} from './index.render'
 import {Input, Button, UserCard} from '../../../common-adapters'
-import {globalColors} from '../../../styles'
+import {globalColors, globalMargins} from '../../../styles'
+
+import type {Props} from '.'
 
 type State = {usernameOrEmail: string}
 
-class UsernameOrEmailRender extends Component<void, Props, State> {
+class UsernameOrEmail extends Component<void, Props, State> {
   state: State
 
   constructor(props: Props) {
@@ -31,21 +32,22 @@ class UsernameOrEmailRender extends Component<void, Props, State> {
       <Container
         style={stylesContainer}
         outerStyle={{backgroundColor: globalColors.white}}
-        onBack={() => this.props.onBack()}
+        onBack={this.props.onBack}
       >
-        <UserCard style={stylesCard} outerStyle={stylesOuterCard}>
+        <UserCard style={stylesCard}>
           <Input
             autoFocus={true}
             style={stylesInput}
             hintText="Username or email"
-            onEnterKeyDown={() => this.onSubmit()}
+            floatingHintTextOverride="Username or email"
             onChangeText={text => this.onChange(text)}
+            onEnterKeyDown={() => this.onSubmit()}
             value={this.state.usernameOrEmail}
           />
           <Button
+            fullWidth={true}
             label="Continue"
             type="Primary"
-            style={{alignSelf: 'center'}}
             onClick={() => this.onSubmit()}
             enabled={this.state.usernameOrEmail}
             waiting={this.props.waitingForResponse}
@@ -58,17 +60,13 @@ class UsernameOrEmailRender extends Component<void, Props, State> {
 
 const stylesContainer = {
   flex: 1,
-  alignItems: 'center',
-  justifyContent: 'center',
 }
 const stylesInput = {
-  marginBottom: 48,
-}
-const stylesOuterCard = {
-  marginTop: 40,
+  flexGrow: 1,
+  marginBottom: globalMargins.small,
 }
 const stylesCard = {
   alignItems: 'stretch',
 }
 
-export default UsernameOrEmailRender
+export default UsernameOrEmail
