+++ shared/login/signup/device-name.js
@@ -1,7 +1,7 @@
 // @flow
 import * as signupActions from '../../actions/signup'
 import React, {Component} from 'react'
-import Render from '../register/set-public-name/index.render'
+import Render from '../register/set-public-name'
 import {bindActionCreators} from 'redux'
 import {connect} from 'react-redux'
